FEAT | Blocked users + style improvements

+++ litrevu/views.py
@@ -1,14 +1,19 @@
-from django.contrib.auth import get_user_model, login
+from itertools import chain
+
+from django.contrib import messages
+from django.contrib.auth import login
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.db.models import CharField, F, Q, Subquery, Value
 from django.shortcuts import get_object_or_404, redirect, render
 
 from litrevu.forms import (
-    FollowUserForm,
     RegistrationForm,
     ReviewForm,
-    TicketForm
+    TicketForm,
+    UserActionForm
 )
-from litrevu.models import Review, Ticket, UserFollows
+from litrevu.models import Review, Ticket, UserBlocked, UserFollows
 
 
 def home(request):
@@ -20,51 +25,53 @@ def register(request):
         form = RegistrationForm(request.POST)
         if form.is_valid():
             user = form.save()
-
             login(request, user)
             return redirect("feed")
+        else:
+            messages.error(request, "Passwords do not match")
     else:
         form = RegistrationForm()
 
     return render(request, "registration/register.html", {"form": form})
 
 
+ID = int
+
+
 @login_required
 def feed(request):
-    # Users followed by current user
-    following_users = UserFollows.objects.filter(user=request.user).values_list(
+    followed_user_ids = UserFollows.objects.filter(user=request.user).values_list(
         "followed_user", flat=True
     )
 
-    # Followed users tickets + User tickets
-    tickets = (
-        Ticket.objects.filter(user__in=following_users)
-        .union(Ticket.objects.filter(user=request.user))
-        .order_by("-time_created")
+    tickets = Ticket.objects.filter(
+        Q(user_id__in=Subquery(followed_user_ids)) | Q(user=request.user)
+    ).annotate(
+        content_type=Value("TICKET", CharField()), ticket_username=F("user__username")
     )
 
-    # Reviews to user tickets
-    my_ticket_ids = Ticket.objects.filter(user=request.user).values_list(
+    user_tickets_subquery = Ticket.objects.filter(user=request.user).values_list(
         "id", flat=True
     )
 
-    reviews_for_my_tickets = Review.objects.filter(ticket__in=my_ticket_ids).order_by(
-        "-time_created"
+    reviews = Review.objects.filter(
+        Q(user_id__in=Subquery(followed_user_ids))
+        | Q(user=request.user)
+        | Q(ticket_id__in=Subquery(user_tickets_subquery))
+    ).annotate(
+        content_type=Value("REVIEW", CharField()),
+        ticket_title=F("ticket__title"),
+        ticket_username=F("ticket__user__username"),
+        review_username=F("user__username"),
     )
 
-    # Reviews by user and followed users
-    reviews = (
-        Review.objects.filter(user__in=following_users)
-        .union(Review.objects.filter(user=request.user))
-        .order_by("-time_created")
+    posts = sorted(
+        chain(reviews, tickets),
+        key=lambda post: post.time_created,
+        reverse=True,
     )
 
-    context = {
-        "tickets": tickets,
-        "reviews": reviews,
-        "reviews_for_my_tickets": reviews_for_my_tickets,
-    }
-    return render(request, "feed.html", context)
+    return render(request, "feed.html", context={"posts": posts})
 
 
 def create_ticket(request):
@@ -126,40 +133,87 @@ def create_review(request, ticket_id):
 
 
 @login_required
-def follow_view(request):
+def social_view(request):
     if request.method == "POST":
-        form = FollowUserForm(request.POST)
+        current_user = request.user
+        form = UserActionForm(request.POST)
         if form.is_valid():
-            username = form.cleaned_data["user_to_follow"]
-            try:
-                followed_user = get_user_model().objects.get(username=username)
-                UserFollows.objects.get_or_create(
-                    user=request.user, followed_user=followed_user
-                )
-            except get_user_model().DoesNotExist:
-                # Gérer le cas où l'utilisateur n'existe pas
-                pass
+            action = form.cleaned_data["action"]
+            target_username = form.cleaned_data["user_to_act_on"]
+            target_user = User.objects.filter(username=target_username).first()
+
+            if target_user:
+                if action == "follow":
+                    user_is_blocking = UserBlocked.objects.filter(
+                        user=current_user, blocked_user=target_user
+                    ).exists()
+                    user_is_blocked = UserBlocked.objects.filter(
+                        user=target_user, blocked_user=current_user
+                    ).exists()
+
+                    if user_is_blocked:
+                        messages.error(
+                            request, "Sorry but you have been blocked by this user!"
+                        )
+                    elif user_is_blocking:
+                        messages.error(
+                            request,
+                            "You need to unblock this user first!",
+                        )
+                    else:
+                        UserFollows.objects.get_or_create(
+                            user=current_user, followed_user=target_user
+                        )
+                        messages.success(request, "User added successfully.")
+
+                elif action == "block":
+                    UserBlocked.objects.get_or_create(
+                        user=request.user, blocked_user=target_user
+                    )
+                    UserFollows.objects.filter(
+                        Q(user=current_user, followed_user=target_user)
+                        | Q(user=target_user, followed_user=current_user)
+                    ).delete()
+                    messages.success(request, "User blocked successfully.")
+            else:
+                messages.error(request, "User not found")
             return redirect("subscriptions")
     else:
-        form = FollowUserForm()
+        form = UserActionForm()
 
     following = UserFollows.objects.filter(user=request.user)
     followers = UserFollows.objects.filter(followed_user=request.user)
-
+    blocked = UserBlocked.objects.filter(user=request.user)
     return render(
         request,
         "follow_view.html",
-        {"form": form, "following": following, "followers": followers},
+        {
+            "form": form,
+            "following": following,
+            "followers": followers,
+            "blocked": blocked,
+        },
     )
 
 
 @login_required
 def unfollow_view(request, user_id):
-    try:
-        followed_user = get_user_model().objects.get(pk=user_id)
+    followed_user = User.objects.filter(pk=user_id).first()
+    if followed_user:
+        followed_user = User.objects.get(pk=user_id)
         UserFollows.objects.filter(
             user=request.user, followed_user=followed_user
         ).delete()
-    except get_user_model().DoesNotExist:
-        pass
+
+    return redirect("subscriptions")
+
+
+@login_required
+def unblock_view(request, user_id):
+    blocked_user = User.objects.filter(pk=user_id).first()
+    if blocked_user:
+        UserBlocked.objects.filter(
+            user=request.user, blocked_user=blocked_user
+        ).delete()
+
     return redirect("subscriptions")
